Add basic folder and file search functions
To search through folders for files, use os.listdir to return a list of 
items in the provided folder. Iterate through these items and if their 
full path is a folder, skip to next item, or search the file for the 
word.

Use a context manager to search through the files. The find method will 
return an index value if the substring arguement is found, else it 
returns -1.