Add recursion to folder search function
Currently the search_folders function is not looking within folders. If 
the current item is a folder, call the search_folders function 
recursively to enter each folder and their files.

If the file is not utf-8 encoded, the context manager will return 
UnicodeDecodeError. Update it to instead read in binary mode and decode 
to utf-8 with ignore mode.