Create fibonacci function
Use a generator method to build a function that returns the next 
fibonacci number each time the function is run.

These methods use co-routines to 'pause' a function, return a value that 
can be processed, and then 'resume' if required. This has the advantage 
of not requiring data to be held in memory before being processed.

Tuple unpacking can also be used to assign data values.