Create Transaction class
A Transaction class should be responsible for creating instances of 
itself. It should have a static function that knows that structure of a 
transactions and the data types it contains.

This will allow Transaction objects to be created using the static 
function without having to create an instance first. At this point data 
type should be converted if neccessary.