Update average calculation using generator expression
The mean method from the statistics module accepts an iterator which 
means a generator expression can be used in place of list comprehension. 
The generator expression will generate values 'just in time', like a 
generator function, and will not create a list object.

This can be more efficient when processing large volumes of data. The 
generator expression is constructed in the same way as list 
comprehension but cannot be indexed over.